refactor(shadervalidate): split packvalidators.cpp by authoring model
The remediation's own confirmation pass caught this: the lint-parity work
took packvalidators.cpp from 1235 to 1421 lines, and the file was already
past the 1150 hard ceiling. Worse, this PR is what put it there — it was
1048 and compliant before, so the grandfather clause never applied and every
line added to it was a finding compounding an earlier one.

Split at the seam the file already had. The three public functions share no
file-local state, and packvalidators.h already declared exactly these three,
so no caller changes:

  packvalidator_overlay.cpp    438  validatePack
  packvalidator_animation.cpp  651  validateAnimationPack, plus the two
                                    compositor-bake helpers whose only
                                    caller it is
  packvalidator_surface.cpp    419  validateSurfacePack

All three are back under the 1000-line target, not merely under the ceiling.

By family rather than by lint kind, deliberately. Each arm mirrors ONE
runtime's assembly order and that runtime's own set of silently-coerced
metadata fields, which is what makes them readable as three narratives; the
lint kinds that genuinely generalise belong in packvalidatorcommon instead,
and moving them there is a separate change that would rewrite diagnostic
text the tests pin.

Code moved verbatim, includes narrowed to what each arm uses. No behaviour
change intended and none observed: all three bundled gates and
test_pack_validators produce the same results as before the split, and the
full suite stays green at 402/402.
