--- src/shadervalidate/packvalidators.cpp
@@ -1,1421 +1,0 @@
-// SPDX-FileCopyrightText: 2026 fuddlesworth
-// SPDX-License-Identifier: GPL-3.0-or-later
-//
-// The three per-authoring-model shader-pack validators — see packvalidators.h.
-// Each reproduces the matching runtime's GLSL assembly and bakes every stage
-// through headless glslang (QShaderBaker, CPU-only). Shared compile/report
-// helpers live in packvalidatorcommon.h.
-
-#include "packvalidators.h"
-
-#include "packvalidatorcommon.h"
-
-#include <PhosphorAnimation/AnimationShaderContract.h>
-#include <PhosphorAnimation/AnimationShaderEffect.h>
-#include <PhosphorAnimation/AnimationShaderRegistry.h>
-#include <PhosphorAnimation/ProfilePaths.h>
-#include <PhosphorRendering/ShaderCompiler.h>
-#include <PhosphorShaders/CustomParamsKey.h>
-#include <PhosphorShaders/ShaderEntryPoint.h>
-#include <PhosphorShaders/ShaderParamPreamble.h>
-#include <PhosphorShaders/ShaderRegistry.h>
-#include <PhosphorSurface/SurfaceShaderContract.h>
-#include <PhosphorSurface/SurfaceShaderEffect.h>
-#include <PhosphorSurface/SurfaceShaderRegistry.h>
-
-#include <QDir>
-#include <QFile>
-#include <QFileInfo>
-#include <QHash>
-#include <QJsonArray>
-#include <QJsonDocument>
-#include <QJsonObject>
-#include <QSet>
-#include <QRegularExpression>
-#include <QString>
-#include <QStringList>
-#include <QTextStream>
-
-#include <rhi/qshader.h>
-
-using PhosphorAnimationShaders::AnimationShaderEffect;
-using PhosphorAnimationShaders::AnimationShaderRegistry;
-using PhosphorRendering::ShaderCompiler;
-using PhosphorShaders::ShaderRegistry;
-using PhosphorSurfaceShaders::SurfaceShaderEffect;
-using PhosphorSurfaceShaders::SurfaceShaderRegistry;
-
-namespace PlasmaZones::ShaderValidate {
-
-// Validate one pack directory. Returns the number of errors found.
-int validatePack(const QString& packDir, QTextStream& out)
-{
-    const QString name = QFileInfo(packDir).fileName();
-
-    QString parseErr;
-    ShaderRegistry::ShaderInfo info = ShaderRegistry::parsePackMetadata(packDir, &parseErr);
-    if (!parseErr.isEmpty()) {
-        out << name << "\n  metadata       ERROR\n    " << parseErr << "\n  → 1 error\n\n";
-        return 1;
-    }
-
-    // Several lints below deliberately read the RAW metadata rather than the
-    // parsed ShaderInfo (parsePackMetadata clamps bufferScale, clears missing
-    // buffer paths, and only sets vertexShaderPath when the file exists, so a
-    // lint reading the parsed struct would silently pass the author error the
-    // runtime hid). Read and parse metadata.json ONCE here and share it, instead
-    // of reopening the same file per lint block.
-    // Checked on both arms: an empty rawRoot silently no-ops EVERY raw-metadata
-    // lint below (the declared-vertexShader check, the duplicate parameter id
-    // check, the buffer name and bufferScale checks), so the pack would report
-    // clean having had those checks skipped rather than passed. parsePackMetadata
-    // above rejects unparseable JSON, so the live path here is an open failure,
-    // but a lint block that can quietly evaluate against nothing is worth
-    // closing whichever way it is reached.
-    QJsonObject rawRoot;
-    {
-        QFile metaFile(QDir(packDir).filePath(QStringLiteral("metadata.json")));
-        if (!metaFile.open(QIODevice::ReadOnly)) {
-            out << "  metadata.json  ERROR\n    cannot read " << metaFile.fileName() << ": " << metaFile.errorString()
-                << "\n";
-            return 1;
-        }
-        QJsonParseError parseError;
-        const QJsonDocument doc = QJsonDocument::fromJson(metaFile.readAll(), &parseError);
-        if (parseError.error != QJsonParseError::NoError || !doc.isObject()) {
-            out << "  metadata.json  ERROR\n    "
-                << (parseError.error != QJsonParseError::NoError ? parseError.errorString()
-                                                                 : QStringLiteral("not a JSON object"))
-                << "\n";
-            return 1;
-        }
-        rawRoot = doc.object();
-    }
-
-    // The `fragmentShader` / `bufferShaders` / `vertexShader` paths come from
-    // the user-editable metadata.json. parsePackMetadata already confines them
-    // to the pack dir (resolveWithinPack), so this re-check is belt-and-braces
-    // defense in depth before each path is opened and fed to glslang — the
-    // same guard the animation and surface validators apply. Zone packs have
-    // no `builtin:` buffer tokens (that resolver is surface-only), so every
-    // path must stay inside the pack. An empty path is left as-is (that stage
-    // is simply absent).
-    const auto confineToPack = [&packDir](QString& path) {
-        return confinePackPathInPlace(packDir, path);
-    };
-    if (!confineToPack(info.sourcePath)) {
-        out << name
-            << "\n  metadata       ERROR\n    fragmentShader path escapes the pack directory (path traversal "
-               "rejected)\n  → 1 error\n\n";
-        return 1;
-    }
-    for (QString& buf : info.bufferShaderPaths) {
-        if (!confineToPack(buf)) {
-            out << name
-                << "\n  metadata       ERROR\n    bufferShaders path escapes the pack directory (path traversal "
-                   "rejected)\n  → 1 error\n\n";
-            return 1;
-        }
-    }
-    if (!confineToPack(info.vertexShaderPath)) {
-        out << name
-            << "\n  metadata       ERROR\n    vertexShader path escapes the pack directory (path traversal rejected)\n "
-               " "
-               "→ 1 error\n\n";
-        return 1;
-    }
-
-    out << name << "  (" << info.parameters.size() << " param" << (info.parameters.size() == 1 ? "" : "s") << ", "
-        << (info.isMultipass ? "multipass" : "single-pass") << ")\n";
-
-    int errors = 0;
-
-    // ── metadata lints ──
-    QStringList lints;
-    QHash<QString, QString> claimedLane; // "pool#slot" → first param id, for collision detection
-    for (const ShaderRegistry::ParameterInfo& p : info.parameters) {
-        if (!kValidParamTypes.contains(p.type)) {
-            lints << QStringLiteral("unknown param type '%1' for '%2'").arg(p.type, p.id);
-        }
-        // An id that isn't a valid GLSL identifier gets no p_ define and no lane
-        // (parseShaderMetadata leaves its slot at -1) — surface the real fault, and
-        // skip the collision check so two such params don't false-collide on "-1".
-        if (!PhosphorShaders::isValidParamId(p.id)) {
-            lints
-                << QStringLiteral("invalid parameter id '%1' (not a GLSL identifier; skipped, no p_ define)").arg(p.id);
-            continue;
-        }
-        const QString laneKey = poolName(p.type) + QStringLiteral("#") + QString::number(p.slot);
-        if (claimedLane.contains(laneKey)) {
-            lints << QStringLiteral("slot collision: '%1' and '%2' both map to %3 lane %4")
-                         .arg(claimedLane.value(laneKey), p.id, poolName(p.type))
-                         .arg(p.slot);
-        } else {
-            claimedLane.insert(laneKey, p.id);
-        }
-        // Mirror the pool budgets the runtime binds against: a slot past its
-        // pool's ceiling gets no p_ define and no upload, so the shader
-        // compiles against an undefined identifier.
-        const int budget = p.type == QLatin1String("color") ? PhosphorShaders::CustomColors::kColorCount
-            : p.type == QLatin1String("image")              ? PhosphorShaders::kMaxImageSlots
-                                                            : PhosphorShaders::CustomParams::kFlatSlotCount;
-        if (p.slot >= budget) {
-            lints << QStringLiteral("slot %1 past the %2-pool budget of %3 for '%4' (will not bind)")
-                         .arg(p.slot)
-                         .arg(poolName(p.type))
-                         .arg(budget)
-                         .arg(p.id);
-        }
-    }
-    // Per-pool count overflow, independent of how slots were assigned. The
-    // per-param `slot >= budget` lint above only fires for params that LAND on
-    // an out-of-range lane; it misses the case where more params request a pool
-    // than it has lanes but they pile onto in-budget slots (explicit collisions,
-    // or a future change to the auto-assigner). Count the lane-consuming params
-    // per pool (valid id, that pool's type) and flag a pool that is oversubscribed
-    // as a whole, so the author sees the real cause rather than only the symptom.
-    {
-        int scalarCount = 0, colorCount = 0, imageCount = 0;
-        for (const ShaderRegistry::ParameterInfo& p : info.parameters) {
-            if (!PhosphorShaders::isValidParamId(p.id) || !kValidParamTypes.contains(p.type)) {
-                continue;
-            }
-            if (p.type == QLatin1String("color")) {
-                ++colorCount;
-            } else if (p.type == QLatin1String("image")) {
-                ++imageCount;
-            } else {
-                ++scalarCount;
-            }
-        }
-        const auto flagPool = [&lints](const QString& pool, int count, int budget) {
-            if (count > budget) {
-                lints << QStringLiteral(
-                             "%1 pool oversubscribed: %2 parameters for %3 lanes (the last %4 will not bind)")
-                             .arg(pool)
-                             .arg(count)
-                             .arg(budget)
-                             .arg(count - budget);
-            }
-        };
-        flagPool(poolName(QStringLiteral("float")), scalarCount, PhosphorShaders::CustomParams::kFlatSlotCount);
-        flagPool(poolName(QStringLiteral("color")), colorCount, PhosphorShaders::CustomColors::kColorCount);
-        flagPool(poolName(QStringLiteral("image")), imageCount, PhosphorShaders::kMaxImageSlots);
-    }
-    // Buffer-pass + bufferScale lints check the RAW metadata, not the parsed
-    // ShaderInfo: parseShaderMetadata clamps bufferScale into [0.125, 1.0] and
-    // clears bufferShaderPaths when a declared buffer is missing, so a lint reading
-    // the parsed values would silently pass an author error the runtime hid.
-    //
-    // The GATE is raw for the same reason, and it has to be: parseShaderMetadata
-    // sets info.isMultipass = false as its fail-closed response to an
-    // unresolvable buffer entry, which is precisely the defect these lints
-    // report. Gating on the parsed flag meant the block was skipped for every
-    // pack that needed it and ran only for packs that were already fine.
-    const bool authorDeclaredMultipass = rawRoot.value(QLatin1String("multipass")).toBool(false)
-        || !rawRoot.value(QLatin1String("bufferShaders")).toArray().isEmpty()
-        || !rawRoot.value(QLatin1String("bufferShader")).toString().isEmpty();
-    if (authorDeclaredMultipass) {
-        const QJsonObject& root = rawRoot;
-
-        const QJsonArray declared = root.value(QLatin1String("bufferShaders")).toArray();
-        QStringList bufferNames;
-        if (declared.isEmpty()) {
-            // The implicit fallback keys off the RAW array being absent, the
-            // way parseShaderMetadata's bufferShadersDeclared does. Keying it
-            // off the non-empty SUBSET instead made `"bufferShaders": [""]`
-            // lint a buffer.frag the runtime never looks at, while saying
-            // nothing about the empty entry that is the actual defect.
-            bufferNames << root.value(QLatin1String("bufferShader")).toString(QStringLiteral("buffer.frag"));
-        } else {
-            for (const QJsonValue& v : declared) {
-                bufferNames << v.toString();
-            }
-        }
-        for (const QString& bufName : bufferNames) {
-            if (bufName.isEmpty()) {
-                // resolveWithinPack returns empty for an empty name before it
-                // reaches the traversal guard, and the caller treats an empty
-                // resolve as fail-closed: it clears the WHOLE bufferShaderPaths
-                // list and turns multipass off for the entire pack. Silent
-                // without this lint, and the warning the runtime does log
-                // misattributes it to a path escape.
-                lints << QStringLiteral(
-                    "empty bufferShaders entry (resolves empty at load, which drops the whole bufferShaders "
-                    "list and disables multipass for the pack)");
-                continue;
-            }
-            const auto confined = confinedPackPath(packDir, bufName);
-            if (!confined) {
-                lints << QStringLiteral(
-                             "multipass buffer shader path escapes the pack directory: %1 (drops the "
-                             "whole bufferShaders list and disables multipass for the pack)")
-                             .arg(bufName);
-            } else if (!QFile::exists(*confined)) {
-                lints << QStringLiteral(
-                             "multipass buffer shader missing: %1 (drops the whole bufferShaders list "
-                             "and disables multipass for the pack)")
-                             .arg(bufName);
-            }
-        }
-
-        // The runtime caps buffer passes at 4 (parseShaderMetadata's qMin) and
-        // drops the surplus with only a journal warning — exactly the
-        // "runtime hid the author error" class this block lints for.
-        // Compare the RAW declared array against the cap, not the non-empty
-        // subset: parseShaderMetadata iterates qMin(rawSize, kMaxBufferPasses)
-        // and only then skips empties, so ["", "a", "b", "c", "d"] silently
-        // loses "d".
-        if (declared.size() > PhosphorShaders::kMaxBufferPasses) {
-            lints << QStringLiteral("too many buffer shaders: %1 declared, cap is %2 (surplus dropped at load)")
-                         .arg(static_cast<int>(declared.size()))
-                         .arg(PhosphorShaders::kMaxBufferPasses);
-        }
-
-        const double rawScale = root.value(QLatin1String("bufferScale")).toDouble(1.0);
-        if (rawScale < PhosphorShaders::kMinBufferScale || rawScale > PhosphorShaders::kMaxBufferScale) {
-            lints << QStringLiteral("bufferScale out of range [%1, %2]: %3 (clamped at load)")
-                         .arg(PhosphorShaders::kMinBufferScale)
-                         .arg(PhosphorShaders::kMaxBufferScale)
-                         .arg(rawScale);
-        }
-
-        // Wrap/filter vocabulary and alignment, the lints the animation and
-        // surface arms already carry. This arm needs them MOST: normalizeWrapMode
-        // and normalizeFilterMode map an unrecognised token to clamp/linear with
-        // no warning at any layer, and parseShaderMetadata pads a short array
-        // with the single-value default and trims a long one, also silently. The
-        // sibling runtimes at least log.
-        const auto lintTokens = [&lints](const QJsonArray& arr, const QString& field, bool wrap) {
-            for (const QJsonValue& v : arr) {
-                const QString tok = v.toString();
-                const bool ok =
-                    wrap ? PhosphorShaders::isValidWrapToken(tok) : PhosphorShaders::isValidFilterToken(tok);
-                if (!tok.isEmpty() && !ok) {
-                    lints << QStringLiteral("%1 value '%2' not in vocabulary (coerced at load)").arg(field, tok);
-                }
-            }
-        };
-        const QJsonArray wrapsArr = root.value(QLatin1String("bufferWraps")).toArray();
-        const QJsonArray filtersArr = root.value(QLatin1String("bufferFilters")).toArray();
-        lintTokens(wrapsArr, QStringLiteral("bufferWraps"), true);
-        lintTokens(filtersArr, QStringLiteral("bufferFilters"), false);
-        const auto lintLength = [&lints, &declared](const QJsonArray& arr, const QString& field) {
-            if (!arr.isEmpty() && arr.size() != declared.size()) {
-                lints << QStringLiteral(
-                             "%1 has %2 entries for %3 buffer shaders (aligned positionally; "
-                             "surplus dropped and missing entries fall back at load)")
-                             .arg(field)
-                             .arg(static_cast<int>(arr.size()))
-                             .arg(static_cast<int>(declared.size()));
-            }
-        };
-        lintLength(wrapsArr, QStringLiteral("bufferWraps"));
-        lintLength(filtersArr, QStringLiteral("bufferFilters"));
-        const QString singleWrap = root.value(QLatin1String("bufferWrap")).toString();
-        if (!singleWrap.isEmpty() && !PhosphorShaders::isValidWrapToken(singleWrap)) {
-            lints << QStringLiteral("bufferWrap value '%1' not in vocabulary (coerced at load)").arg(singleWrap);
-        }
-        const QString singleFilter = root.value(QLatin1String("bufferFilter")).toString();
-        if (!singleFilter.isEmpty() && !PhosphorShaders::isValidFilterToken(singleFilter)) {
-            lints << QStringLiteral("bufferFilter value '%1' not in vocabulary (coerced at load)").arg(singleFilter);
-        }
-    }
-    if (!QFile::exists(info.sourcePath)) {
-        lints << QStringLiteral("fragment shader missing: %1").arg(QFileInfo(info.sourcePath).fileName());
-    }
-    // A declared-but-absent vertexShader silently falls back to the shared
-    // zone.vert below, hiding the author's typo. Read the RAW metadata (shared
-    // rawRoot above): unlike the animation and surface parsers,
-    // ShaderRegistry::parsePackMetadata only assigns vertexShaderPath when the
-    // file EXISTS, so linting the parsed struct could never fire.
-    {
-        const QString declaredVert = rawRoot.value(QLatin1String("vertexShader")).toString();
-        if (!declaredVert.isEmpty() && !QFile::exists(QDir(packDir).filePath(declaredVert))) {
-            lints << QStringLiteral("vertex shader missing: %1").arg(declaredVert);
-        }
-        // A custom-named vertexShader is silently ignored by the zone
-        // runtime, which only ever loads a file named `zone.vert` (see the
-        // vertex-resolution note below). Warn so the author is not misled by
-        // a green validator into thinking their custom vertex stage runs.
-        if (!declaredVert.isEmpty() && QFileInfo(declaredVert).fileName() != QLatin1String("zone.vert")) {
-            lints << QStringLiteral(
-                         "vertexShader '%1' is ignored at load: the zone runtime only uses a file named "
-                         "zone.vert (sibling, else shared). Rename it to zone.vert or drop the declaration.")
-                         .arg(declaredVert);
-        }
-        // Duplicate ids must be linted from the RAW array: the parser
-        // dedupes (first declaration wins), so the parsed struct can
-        // never show the author their repeated id.
-        QSet<QString> rawIds;
-        const QJsonArray rawParams = rawRoot.value(QLatin1String("parameters")).toArray();
-        for (const QJsonValue& v : rawParams) {
-            const QString rawId = v.toObject().value(QLatin1String("id")).toString();
-            if (rawId.isEmpty()) {
-                continue;
-            }
-            if (rawIds.contains(rawId)) {
-                lints << QStringLiteral("duplicate parameter id '%1' (first declaration wins at load)").arg(rawId);
-            } else {
-                rawIds.insert(rawId);
-            }
-        }
-    }
-
-    if (lints.isEmpty()) {
-        out << "  metadata       OK\n";
-    } else {
-        out << "  metadata       ERROR\n";
-        for (const QString& l : lints) {
-            out << "    " << l << "\n";
-            ++errors;
-        }
-    }
-
-    // ── stage compiles (reproduce the runtime assembly) ──
-    const QString packsRoot = QFileInfo(packDir).absolutePath();
-    const QStringList includePaths = {packsRoot + QStringLiteral("/shared"), packsRoot};
-    const QString preamble = ShaderRegistry::paramPreamble(info);
-
-    if (QFile::exists(info.sourcePath)) {
-        // Label by the actual fragment filename — parsePackMetadata honours a
-        // custom `fragmentShader` field (default effect.frag), so the stage label
-        // tracks the real file rather than assuming effect.frag.
-        errors += compileStage(out, QFileInfo(info.sourcePath).fileName(), info.sourcePath, QShader::FragmentStage,
-                               includePaths, /*useScaffold=*/true, preamble, info);
-    }
-    // Only multipass packs bake buffer passes — parseShaderMetadata clears
-    // buffer state for single-pass packs and only keeps an implicit
-    // buffer.frag that exists on disk, and the runtime (bakeBufferShaders)
-    // ignores buffer paths unless isMultipass, so the validator gates
-    // identically.
-    if (info.isMultipass) {
-        for (const QString& buf : info.bufferShaderPaths) {
-            if (QFile::exists(buf)) {
-                errors += compileStage(out, QFileInfo(buf).fileName(), buf, QShader::FragmentStage, includePaths,
-                                       /*useScaffold=*/false, QString(), info);
-            }
-        }
-    }
-    // Vertex: resolve EXACTLY as the zone runtime does (resolveZoneVertexPath in
-    // zoneshadernoderhi.h) rather than honouring `info.vertexShaderPath`. The
-    // zone runtime — both ZoneShaderItem and the daemon warm bake — only ever
-    // looks for a file literally named `zone.vert` (the pack's sibling, else the
-    // shared copy) and NEVER compiles a custom-named `vertexShader` declaration,
-    // unlike the animation and surface runtimes. Baking `info.vertexShaderPath`
-    // here would compile a stage the runtime never touches, so a pack with a
-    // broken custom vert would pass CI while its real (shared) vertex stage went
-    // unchecked. So resolve the sibling zone.vert, else the shared zone.vert.
-    QString vertPath = QFileInfo(info.sourcePath).absolutePath() + QStringLiteral("/zone.vert");
-    if (!QFile::exists(vertPath)) {
-        vertPath = packsRoot + QStringLiteral("/shared/zone.vert");
-    }
-    if (QFile::exists(vertPath)) {
-        errors += compileStage(out, QFileInfo(vertPath).fileName(), vertPath, QShader::VertexStage, includePaths,
-                               /*useScaffold=*/false, QString(), info);
-    }
-
-    if (errors == 0) {
-        out << "  → OK\n\n";
-    } else {
-        out << "  → " << errors << (errors == 1 ? " error\n\n" : " errors\n\n");
-    }
-    return errors;
-}
-
-// Which of the compositor-only shared-helper samplers the (include-expanded)
-// fragment source references. shared/old_content.glsl,
-// shared/desktop_transition.glsl and shared/strip_transition.glsl declare
-// these binding-less, which the strict
-// SPIR-V bake rejects — so when a daemon-eligible pack's compile fails with
-// the binding-less-sampler diagnostic AND its source pulls one of these in,
-// the likely root cause is a metadata bug (missing compositor-only
-// appliesTo), not a GLSL bug, and deserves a hint that says so. The source
-// scan alone is NOT a lint: daemon-capable packs legitimately reference these
-// samplers inside `#ifdef PLASMAZONES_KWIN` branches, which the preprocessor
-// strips before they can fail the bake — hence the hint is gated on the
-// compile actually failing with glslang's binding diagnostic (which does not
-// echo the offending identifier, so the source scan supplies the name). The
-// combination is still heuristic — a pack whose OWN binding-less sampler
-// fails while a guarded compositor-sampler reference coexists would draw the
-// hint spuriously — which is why this stays an advisory hint appended to the
-// real compile error, never an error of its own.
-static QStringList compositorOnlySamplersUsed(const QString& expandedSource)
-{
-    struct SamplerMatcher
-    {
-        QString name;
-        // Word-boundary match so a longer identifier (e.g.
-        // "uOldWindowFallback") can't count as a use of the sampler.
-        QRegularExpression wordRe;
-    };
-    static const auto kCompositorOnlySamplers = [] {
-        QList<SamplerMatcher> matchers;
-        for (const QString& name : {QStringLiteral("uOldWindow"), QStringLiteral("uFromDesktop"),
-                                    QStringLiteral("uToDesktop"), QStringLiteral("uStrip")}) {
-            matchers.append({name, QRegularExpression(QStringLiteral("\\b") + name + QStringLiteral("\\b"))});
-        }
-        return matchers;
-    }();
-    QStringList used;
-    for (const SamplerMatcher& sampler : kCompositorOnlySamplers) {
-        if (expandedSource.contains(sampler.wordRe)) {
-            used << sampler.name;
-        }
-    }
-    return used;
-}
-
-// Validate one ANIMATION pack directory (data/animations/*). Reproduces the
-// animation runtime's fragment assembly on the daemon (Qt-RHI) path — the entry
-// scaffold (pTransition / pIn+pOut, or a pass-through main()), the generated
-// p_<id> preamble, and include expansion — then bakes through headless glslang.
-// Returns the number of errors found.
-//
-// The kwin-effect classic-GL path (`#define PLASMAZONES_KWIN`, default-block
-// uniforms) is NOT baked here: QShaderBaker compiles Vulkan-dialect GLSL and
-// rejects default-block uniforms, so the kwin branch needs a separate
-// OpenGL-target compiler. Compositor-only packs (desktop / geometry / move /
-// strip / tab classes — see shaderEffectIsCompositorOnly) are authored against that kwin
-// dialect directly and never run on the daemon, so their stages are baked out
-// of process by bakeCompositorStage instead, through glslang in DEFAULT mode.
-// test_animation_shader_kwin_bake remains the additional driver-level check.
-// Daemon-capable packs get the full stage compile below.
-
-// COVERAGE BOUNDARY, so this is not read as more than it is: the source is
-// handed to glslang with the pack's own `#version 450` intact, while KWin's
-// generateCustomShader recompiles at the GL context's core version (140 on the
-// 6.7 path this dialect exists for). So a construct legal at 450 and illegal
-// at 140 still passes here and still fails live — which is the class the ARB
-// extension block in kwinDefineBlock was added to fix. Rewriting the version
-// line before the bake would close it, and would need a trial run against
-// every bundled compositor-only pack first, since it can only make the gate
-// stricter. What this gate does cover is everything a wrong version would not
-// have caught anyway: undeclared identifiers, type errors, bad swizzles, a
-// p_<id> the preamble never emitted.
-//
-// Assemble one stage of a COMPOSITOR-ONLY animation pack exactly as the
-// kwin-effect does (entry scaffold for the fragment, include expansion, the
-// p_<id> preamble, then the shared KWin define block after #version) and
-// compile it through glslangValidator. Returns 1 on failure, 0 on success.
-//
-// The splice order matters and mirrors the runtime: each spliceAfterVersion
-// lands its block immediately below #version, so splicing the preamble first
-// and the define block second leaves the define block ABOVE the preamble,
-// which is what the compositor produces (the preamble, then
-// injectKwinDefineAfterVersion).
-static int bakeCompositorStage(QTextStream& out, const QString& packDir,
-                               const PhosphorAnimationShaders::AnimationShaderEffect& eff, const QString& path,
-                               const QString& label, const QString& stage, bool scaffold)
-{
-    // The absent-stage bail comes FIRST so the two arms agree on the same
-    // input: a pack whose declared stage file is missing is already linted by
-    // the caller, and reporting a missing tool for it would describe the
-    // machine rather than the pack, on a machine without glslang only.
-    if (!QFile::exists(path)) {
-        return 0; // an absent stage is already linted by the caller
-    }
-    const QString tool = glslangValidatorPath();
-    if (tool.isEmpty()) {
-        // Hard failure rather than a skip. The point of this gate is that a
-        // pack cannot reach a release uncompiled, and a quiet "no tool, no
-        // coverage" degrade is how 35 packs went unchecked in the first place.
-        // Only compositor-only packs reach here, so a pack tree containing
-        // none of them still validates on a machine without glslang.
-        out << "  " << label.leftJustified(15)
-            << "ERROR\n    neither glslangValidator nor glslang found on PATH. One of them is required to "
-               "compile compositor-only packs (install the glslang package)\n";
-        return 1;
-    }
-    QFile f(path);
-    if (!f.open(QIODevice::ReadOnly | QIODevice::Text)) {
-        out << "  " << label.leftJustified(15) << "ERROR\n    cannot read " << path << "\n";
-        return 1;
-    }
-    const QString raw = QString::fromUtf8(f.readAll());
-    const QString assembled = scaffold
-        ? PhosphorShaders::assembleEntryPoint(raw, AnimationShaderRegistry::animationEntryPrologue(),
-                                              AnimationShaderRegistry::animationEntryCandidates())
-        : raw;
-    const QStringList includePaths = {QFileInfo(packDir).absolutePath() + QStringLiteral("/shared")};
-    QString err;
-    QString src = ShaderCompiler::expandSource(assembled, QFileInfo(path).absolutePath(), includePaths, &err);
-    if (src.isEmpty()) {
-        out << "  " << label.leftJustified(15) << "ERROR\n    include expansion failed: " << err << "\n";
-        return 1;
-    }
-    src = PhosphorShaders::spliceAfterVersion(src, AnimationShaderRegistry::paramPreamble(eff));
-    src = PhosphorShaders::spliceAfterVersion(src, PhosphorShaders::kwinDefineBlock());
-    return reportCompositorCompile(out, label, stage, src, tool);
-}
-
-int validateAnimationPack(const QString& packDir, QTextStream& out)
-{
-    const QString name = QFileInfo(packDir).fileName();
-
-    QFile metaFile(QDir(packDir).filePath(QStringLiteral("metadata.json")));
-    if (!metaFile.open(QIODevice::ReadOnly)) {
-        out << name << "\n  metadata       ERROR\n    cannot read metadata.json\n  → 1 error\n\n";
-        return 1;
-    }
-    QJsonParseError perr{};
-    const QJsonDocument doc = QJsonDocument::fromJson(metaFile.readAll(), &perr);
-    if (doc.isNull() || !doc.isObject()) {
-        out << name << "\n  metadata       ERROR\n    invalid JSON: " << perr.errorString() << "\n  → 1 error\n\n";
-        return 1;
-    }
-
-    AnimationShaderEffect eff = AnimationShaderEffect::fromJson(doc.object());
-    eff.sourceDir = QDir(packDir).absolutePath();
-    // Mirror AnimationShaderRegistry::parseEffect: the fragment path comes from
-    // the metadata `fragmentShader` field, resolved relative to the pack dir and
-    // confined to it (the metadata is user-editable, so a `../…`/absolute path
-    // must be rejected, not opened and fed to glslang — same guard as
-    // validateSurfacePack). The runtime does NOT default to effect.frag — a pack
-    // that omits the field is unreachable, so an empty path stays empty and is
-    // caught by isValid() below.
-    if (!eff.fragmentShaderPath.isEmpty()) {
-        const auto confined = confinedPackPath(packDir, eff.fragmentShaderPath);
-        if (!confined) {
-            out << name
-                << "\n  metadata       ERROR\n    fragmentShader path escapes the pack directory (path traversal "
-                   "rejected)\n  → 1 error\n\n";
-            return 1;
-        }
-        eff.fragmentShaderPath = *confined;
-    }
-    // The optional `vertexShader` path is user-editable metadata too, so it
-    // gets the same traversal guard as the fragment before the stage bake
-    // below ever opens it.
-    if (!eff.vertexShaderPath.isEmpty()) {
-        const auto confinedVert = confinedPackPath(packDir, eff.vertexShaderPath);
-        if (!confinedVert) {
-            out << name
-                << "\n  metadata       ERROR\n    vertexShader path escapes the pack directory (path traversal "
-                   "rejected)\n  → 1 error\n\n";
-            return 1;
-        }
-        eff.vertexShaderPath = *confinedVert;
-    }
-    if (!eff.isValid()) {
-        out << name << "\n  metadata       ERROR\n    missing required field (id / fragmentShader)\n  → 1 error\n\n";
-        return 1;
-    }
-    const QString fragLabel = QFileInfo(eff.fragmentShaderPath).fileName();
-
-    out << name << "  (" << eff.parameters.size() << " param" << (eff.parameters.size() == 1 ? "" : "s") << ", "
-        << eff.textures.size() << " texture" << (eff.textures.size() == 1 ? "" : "s") << ")\n";
-
-    int errors = 0;
-
-    // ── metadata lints ──
-    // Animation params are always auto-slot (declaration order per pool), so
-    // there are no explicit-slot collisions to detect; lint the param types and
-    // ids that gate whether a p_<id> define is emitted at all.
-    static const QStringList kAnimParamTypes = {QStringLiteral("float"), QStringLiteral("int"), QStringLiteral("bool"),
-                                                QStringLiteral("color")};
-    QStringList lints;
-    for (const AnimationShaderEffect::ParameterInfo& p : eff.parameters) {
-        if (!kAnimParamTypes.contains(p.type)) {
-            lints << QStringLiteral(
-                         "unknown param type '%1' for '%2' (animation params are float/int/bool/color; "
-                         "images are textures, not params)")
-                         .arg(p.type, p.id);
-        }
-        if (!PhosphorShaders::isValidParamId(p.id)) {
-            lints
-                << QStringLiteral("invalid parameter id '%1' (not a GLSL identifier; skipped, no p_ define)").arg(p.id);
-        }
-    }
-    // Lint appliesTo tokens from the RAW metadata (fromJson silently drops
-    // unknown tokens with only a runtime journal warning). A typo matters
-    // doubly here: a compositor-only pack whose every token is misspelled
-    // degrades to universal, escapes the compositor-only skip below, and
-    // fails the daemon stage compile with an opaque GLSL error instead of
-    // a metadata diagnostic. Token handling mirrors fromJson exactly:
-    // trimmed comparison against the ProfilePaths vocabulary constants,
-    // empty / whitespace-only tokens skipped silently.
-    {
-        namespace PP = PhosphorAnimation::ProfilePaths;
-        // The exported vocabulary, not a copy: the lint and fromJson's
-        // acceptance test must never disagree about what a valid token is.
-        static const QStringList kAnimAppliesToTokens = PP::allEventClassTokens();
-        const QJsonValue appliesToValue = doc.object().value(QLatin1String("appliesTo"));
-        // A present-but-non-array appliesTo (e.g. a bare string) is ignored
-        // wholesale by fromJson's .toArray() and the pack silently becomes
-        // universal — a plausible author typo worth a shape diagnostic.
-        if (!appliesToValue.isUndefined() && !appliesToValue.isArray()) {
-            lints << QStringLiteral("appliesTo must be an array of tokens (ignored at load; pack becomes universal)");
-        }
-        // An explicit empty array (`"appliesTo": []`) draws NO lint, and that
-        // is deliberate rather than an oversight — do not "even it up" with
-        // the two cases below. Those two are TYPOS: the author wrote a
-        // constraint, it validated down to nothing, and the pack silently
-        // became universal, which is the opposite of what they asked for. An
-        // empty array is not a typo; it says "no constraint", which is
-        // exactly what it gets. Linting it would flag a legal spelling of
-        // the documented default, and every universal pack that spells it
-        // out would start failing the CI gate. Pinned by
-        // test_pack_validators::explicitEmptyAppliesToIsNotLinted.
-        const QJsonArray declaredAppliesTo = appliesToValue.toArray();
-        for (const QJsonValue& v : declaredAppliesTo) {
-            const QString token = v.toString().trimmed();
-            if (token.isEmpty()) {
-                // fromJson drops these without even a journal warning, so a
-                // stray "" entry is the one appliesTo typo that leaves no
-                // runtime trace at all — and an array that validates down to
-                // empty makes the pack universal.
-                lints << QStringLiteral(
-                    "empty appliesTo token (dropped at load with no runtime warning; an "
-                    "appliesTo that validates down to empty makes the pack universal)");
-                continue;
-            }
-            if (!kAnimAppliesToTokens.contains(token)) {
-                lints << QStringLiteral(
-                             "unknown appliesTo token '%1' (dropped at load; an appliesTo that "
-                             "validates down to empty makes the pack universal; valid tokens are %2)")
-                             .arg(token, kAnimAppliesToTokens.join(QLatin1Char('/')));
-            }
-        }
-    }
-    // Texture lints mirror parseEffect's parse-time journal warnings. fromJson
-    // silently drops these from the in-memory struct, so read the RAW metadata
-    // textures array (same as parseEffect) rather than eff.textures.
-    const QJsonArray declaredTextures = doc.object().value(QLatin1String("textures")).toArray();
-    if (declaredTextures.size() > PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots) {
-        lints << QStringLiteral("too many textures: %1 declared, cap is %2 (surplus dropped at load)")
-                     .arg(static_cast<int>(declaredTextures.size()))
-                     .arg(PhosphorAnimationShaders::AnimationShaderContract::kMaxUserTextureSlots);
-    }
-    for (const QJsonValue& v : declaredTextures) {
-        const QString texPath = v.toObject().value(QLatin1String("path")).toString();
-        if (texPath.isEmpty()) {
-            lints << QStringLiteral("texture entry with empty `path` (dropped at load)");
-        } else {
-            // Confined and existence-checked like every SHADER path this
-            // validator reads. Without it a typo'd or escaping texture shipped
-            // as a green pack: the registry's traversal guard clears the path
-            // and the sampler falls back to transparent black, so the failure
-            // surfaces at the first live playback rather than in CI.
-            const auto confined = confinedPackPath(packDir, texPath);
-            if (!confined) {
-                lints << QStringLiteral(
-                             "texture path escapes the pack directory: %1 (rejected at load, sampler reads "
-                             "transparent)")
-                             .arg(texPath);
-            } else if (!QFile::exists(*confined)) {
-                lints << QStringLiteral("texture missing: %1 (sampler reads transparent at load)").arg(texPath);
-            }
-        }
-        // Wrap vocabulary lint — read RAW metadata: AnimationShaderEffect::fromJson
-        // silently clears an invalid texture wrap to empty, so a lint over the
-        // parsed struct could never surface an author's typo. Mirrors the
-        // surface validator.
-        const QString wrap = v.toObject().value(QLatin1String("wrap")).toString();
-        if (!wrap.isEmpty() && !PhosphorAnimationShaders::AnimationShaderContract::isValidWrapToken(wrap)) {
-            lints
-                << QStringLiteral("texture wrap not in {clamp,repeat,mirror}: %1 (cleared to clamp at load)").arg(wrap);
-        }
-    }
-    // Multipass buffer lints — read RAW metadata, not the parsed struct:
-    // AnimationShaderEffect::fromJson clamps bufferScale, drops missing buffers,
-    // and silently coerces invalid buffer wrap/filter tokens, so a lint over the
-    // parsed values would hide the exact author errors this block exists to
-    // surface. The animation validator had none of these, unlike its two
-    // siblings, even though the animation parser silently rewrites all of them.
-    if (eff.isMultipass) {
-        const QJsonObject animRoot = doc.object();
-        const QJsonArray declaredBuffers = animRoot.value(QLatin1String("bufferShaders")).toArray();
-        for (const QJsonValue& v : declaredBuffers) {
-            const QString bufName = v.toString();
-            if (bufName.isEmpty()) {
-                // Kept in place at load (the arrays below are positionally
-                // aligned with this one), and an empty entry fails the
-                // load-time existence check, which fail-closes multipass for
-                // the whole pack. Silent without this lint.
-                lints << QStringLiteral(
-                    "empty bufferShaders entry (kept in place for alignment; fails the load-time existence "
-                    "check, which fail-closes multipass for the whole pack)");
-                continue;
-            }
-            const auto confined = confinedPackPath(packDir, bufName);
-            if (!confined) {
-                lints << QStringLiteral("multipass buffer shader path escapes the pack directory: %1").arg(bufName);
-            } else if (!QFile::exists(*confined)) {
-                lints << QStringLiteral("multipass buffer shader missing: %1").arg(bufName);
-            }
-        }
-        if (declaredBuffers.size() > PhosphorAnimationShaders::AnimationShaderContract::kMaxBufferPasses) {
-            lints << QStringLiteral("too many buffer shaders: %1 declared, cap is %2 (surplus dropped at load)")
-                         .arg(static_cast<int>(declaredBuffers.size()))
-                         .arg(PhosphorAnimationShaders::AnimationShaderContract::kMaxBufferPasses);
-        }
-        const double rawScale = animRoot.value(QLatin1String("bufferScale")).toDouble(1.0);
-        if (rawScale < PhosphorAnimationShaders::AnimationShaderEffect::kMinBufferScale
-            || rawScale > PhosphorAnimationShaders::AnimationShaderEffect::kMaxBufferScale) {
-            lints << QStringLiteral("bufferScale out of range [%1, %2]: %3 (clamped at load)")
-                         .arg(PhosphorAnimationShaders::AnimationShaderEffect::kMinBufferScale)
-                         .arg(PhosphorAnimationShaders::AnimationShaderEffect::kMaxBufferScale)
-                         .arg(rawScale);
-        }
-        const auto lintTokens = [&lints](const QJsonArray& arr, const QString& field, bool wrap) {
-            for (const QJsonValue& v : arr) {
-                const QString tok = v.toString();
-                const bool ok = wrap ? PhosphorAnimationShaders::AnimationShaderContract::isValidWrapToken(tok)
-                                     : PhosphorAnimationShaders::AnimationShaderContract::isValidFilterToken(tok);
-                if (!tok.isEmpty() && !ok) {
-                    lints << QStringLiteral("%1 value '%2' not in vocabulary (coerced at load)").arg(field, tok);
-                }
-            }
-        };
-        const QJsonArray wrapsArr = animRoot.value(QLatin1String("bufferWraps")).toArray();
-        const QJsonArray filtersArr = animRoot.value(QLatin1String("bufferFilters")).toArray();
-        lintTokens(wrapsArr, QStringLiteral("bufferWraps"), true);
-        lintTokens(filtersArr, QStringLiteral("bufferFilters"), false);
-        // Both arrays are positionally aligned with bufferShaders and trimmed
-        // to its length at load, so a longer one silently loses its tail and a
-        // shorter one leaves the last buffers on the single-value default.
-        // Same length lint the surface validator applies to its own aligned
-        // arrays.
-        const auto lintLength = [&lints, &declaredBuffers](const QJsonArray& arr, const QString& field) {
-            if (!arr.isEmpty() && arr.size() != declaredBuffers.size()) {
-                lints << QStringLiteral(
-                             "%1 has %2 entries for %3 buffer shaders (aligned positionally; "
-                             "surplus dropped and missing entries fall back at load)")
-                             .arg(field)
-                             .arg(static_cast<int>(arr.size()))
-                             .arg(static_cast<int>(declaredBuffers.size()));
-            }
-        };
-        lintLength(wrapsArr, QStringLiteral("bufferWraps"));
-        lintLength(filtersArr, QStringLiteral("bufferFilters"));
-        const QString singleWrap = animRoot.value(QLatin1String("bufferWrap")).toString();
-        if (!singleWrap.isEmpty() && !PhosphorAnimationShaders::AnimationShaderContract::isValidWrapToken(singleWrap)) {
-            lints << QStringLiteral("bufferWrap value '%1' not in vocabulary (coerced at load)").arg(singleWrap);
-        }
-        const QString singleFilter = animRoot.value(QLatin1String("bufferFilter")).toString();
-        if (!singleFilter.isEmpty()
-            && !PhosphorAnimationShaders::AnimationShaderContract::isValidFilterToken(singleFilter)) {
-            lints << QStringLiteral("bufferFilter value '%1' not in vocabulary (coerced at load)").arg(singleFilter);
-        }
-    }
-    if (!QFile::exists(eff.fragmentShaderPath)) {
-        lints << QStringLiteral("fragment shader missing: %1").arg(fragLabel);
-    }
-    // A declared-but-absent vertexShader would silently fall back to
-    // shared/animation.vert at runtime, hiding the author's typo. Same lint
-    // the surface validator applies.
-    if (!eff.vertexShaderPath.isEmpty() && !QFile::exists(eff.vertexShaderPath)) {
-        lints << QStringLiteral("vertex shader missing: %1").arg(QFileInfo(eff.vertexShaderPath).fileName());
-    }
-    // The two screen-level passes draw ONE fixed full-screen quad through
-    // their own vertex stage, so a vertexShader or a geometryGrid declared by
-    // a desktop / strip pack is loaded, kept, and then never used. Say so
-    // rather than let the author wonder why their vertex stage does nothing.
-    {
-        namespace PP = PhosphorAnimation::ProfilePaths;
-        const bool screenLevel =
-            eff.appliesTo.contains(PP::EventClassDesktop) || eff.appliesTo.contains(PP::EventClassStrip);
-        if (screenLevel && !eff.vertexShaderPath.isEmpty()) {
-            lints << QStringLiteral(
-                "vertexShader is ignored for desktop/strip packs (the pass draws its own "
-                "full-screen quad)");
-        }
-        if (screenLevel && eff.geometryGridSubdivisions > 0) {
-            lints << QStringLiteral(
-                "geometryGrid is ignored for desktop/strip packs (the pass draws its own "
-                "full-screen quad)");
-        }
-    }
-    // geometryGrid is qBound(0, raw, cap) at load, so a negative value
-    // silently becomes 0 — indistinguishable from not declaring it, and the
-    // vertex stage the author wanted never subdivides. The over-cap side
-    // already warns on the journal at load; this is the quiet side.
-    {
-        const QJsonValue gridVal = doc.object().value(QLatin1String("geometryGrid"));
-        if (!gridVal.isUndefined() && gridVal.toInt() < 0) {
-            lints << QStringLiteral("geometryGrid is negative (%1); it clamps to 0 at load, disabling the grid")
-                         .arg(gridVal.toInt());
-        }
-    }
-
-    if (lints.isEmpty()) {
-        out << "  metadata       OK\n";
-    } else {
-        out << "  metadata       ERROR\n";
-        for (const QString& l : lints) {
-            out << "    " << l << "\n";
-            ++errors;
-        }
-    }
-
-    // ── stage compile (reproduce the daemon runtime fragment assembly) ──
-    // Compositor-only packs take the out-of-process glslang bake instead of
-    // the SPIR-V one: their source is kwin classic-GL (default-block uniforms,
-    // unbound samplers) that the strict SPIR-V target rejects by design, and
-    // the daemon never loads them.
-    if (PhosphorAnimationShaders::shaderEffectIsCompositorOnly(eff)) {
-        errors += bakeCompositorStage(out, packDir, eff, eff.fragmentShaderPath, fragLabel, QStringLiteral("frag"),
-                                      /*scaffold=*/true);
-    } else if (QFile::exists(eff.fragmentShaderPath)) {
-        QFile frag(eff.fragmentShaderPath);
-        if (!frag.open(QIODevice::ReadOnly | QIODevice::Text)) {
-            out << "  " << fragLabel.leftJustified(15) << "ERROR\n    cannot read " << eff.fragmentShaderPath << "\n";
-            ++errors;
-        } else {
-            const QString raw = QString::fromUtf8(frag.readAll());
-            // An entry-only pack (pTransition / pIn+pOut) gets a generated main();
-            // a traditional main() pack passes through — exactly as both runtimes do.
-            const QString assembled =
-                PhosphorShaders::assembleEntryPoint(raw, AnimationShaderRegistry::animationEntryPrologue(),
-                                                    AnimationShaderRegistry::animationEntryCandidates());
-            // Animation runtime include paths are `shared`-only (see
-            // surfaceanimator.cpp animIncludePaths, which appends only each
-            // search path's `/shared` subdir), so the animation gate matches it.
-            const QStringList includePaths = {QFileInfo(packDir).absolutePath() + QStringLiteral("/shared")};
-            QString err;
-            const QString expanded = ShaderCompiler::expandSource(
-                assembled, QFileInfo(eff.fragmentShaderPath).absolutePath(), includePaths, &err);
-            if (expanded.isEmpty()) {
-                out << "  " << fragLabel.leftJustified(15) << "ERROR\n    include expansion failed: " << err << "\n";
-                ++errors;
-            } else {
-                const QString spliced =
-                    PhosphorShaders::spliceAfterVersion(expanded, AnimationShaderRegistry::paramPreamble(eff));
-                const ShaderCompiler::Result result = ShaderCompiler::compile(spliced.toUtf8(), QShader::FragmentStage);
-                errors += reportCompile(out, fragLabel, result, declaredParamNames(eff.parameters));
-                // A binding-less-sampler failure in a pack that pulls in the
-                // compositor-only shared helpers is a metadata bug in disguise
-                // — point the author at the real fix instead of leaving the
-                // opaque GLSL error. (glslang's diagnostic does not name the
-                // sampler, so match the failure mode + the source reference.)
-                if (!result.success && result.error.contains(QLatin1String("requires layout(binding"))) {
-                    const QStringList kwinOnly = compositorOnlySamplersUsed(expanded);
-                    if (!kwinOnly.isEmpty()) {
-                        out << "    (hint: " << kwinOnly.join(QStringLiteral(", "))
-                            << " is declared by the compositor-only shared/old_content.glsl / "
-                               "shared/desktop_transition.glsl / shared/strip_transition.glsl helpers; "
-                               "packs that use them outside "
-                               "a PLASMAZONES_KWIN guard must declare a compositor-only appliesTo — "
-                               "omit \"appearance\")\n";
-                    }
-                }
-            }
-        }
-    }
-
-    // ── multipass buffer passes ──
-    // The zone and surface validators have always baked their buffer passes;
-    // the animation validator never did, so a broken buffer shader in an
-    // animation pack passed this gate and failed only at the live daemon.
-    //
-    // Reproduce ShaderNodeRhi::bakeBufferShaders exactly: load, expand
-    // includes, compile. NO entry scaffold (there is no pTransition to wrap)
-    // and — verified against that function — NO p_<id> preamble either, so a
-    // buffer pass that wants the pack's parameters must include
-    // <animation_uniforms.glsl> itself and read customParams directly.
-    // Splicing a preamble the runtime does not splice would be worse than no
-    // coverage: the gate would pass sources that fail live and fail sources
-    // that work.
-    //
-    // Skipped for compositor-only packs for the same reason the fragment
-    // stage is: the strict SPIR-V target rejects their dialect by design.
-    // Unlike the fragment and vertex stages these do NOT take the glslang
-    // bake either, so a compositor-only multipass pack's buffer passes are
-    // compiled nowhere. No bundled pack is both, and closing it means
-    // teaching bakeCompositorStage the per-pass uniform contract, which is a
-    // larger change than the stage bake was.
-    if (eff.isMultipass && !PhosphorAnimationShaders::shaderEffectIsCompositorOnly(eff)) {
-        const QStringList includePaths = {QFileInfo(packDir).absolutePath() + QStringLiteral("/shared")};
-        for (const QString& declaredBuf : eff.bufferShaderPaths) {
-            // fromJson leaves these RELATIVE (unlike the fragment path, which
-            // the block at the top of this function resolves), so resolve
-            // against the pack dir the way the runtime's parseEffect does.
-            // Baking the raw string would silently find nothing and report a
-            // clean pack — which is exactly how this whole stage was missing.
-            if (declaredBuf.isEmpty()) {
-                continue; // already linted above
-            }
-            const auto confinedBuf = confinedPackPath(packDir, declaredBuf);
-            if (!confinedBuf || !QFile::exists(*confinedBuf)) {
-                continue; // escaping and absent buffers are already linted above
-            }
-            const QString buf = *confinedBuf;
-            const QString label = QFileInfo(buf).fileName();
-            QFile bufFile(buf);
-            if (!bufFile.open(QIODevice::ReadOnly | QIODevice::Text)) {
-                out << "  " << label.leftJustified(15) << "ERROR\n    cannot read " << buf << "\n";
-                ++errors;
-                continue;
-            }
-            const QString rawBuf = QString::fromUtf8(bufFile.readAll());
-            QString bufErr;
-            const QString expandedBuf =
-                ShaderCompiler::expandSource(rawBuf, QFileInfo(buf).absolutePath(), includePaths, &bufErr);
-            if (expandedBuf.isEmpty()) {
-                out << "  " << label.leftJustified(15) << "ERROR\n    include expansion failed: " << bufErr << "\n";
-                ++errors;
-                continue;
-            }
-            const ShaderCompiler::Result bufResult =
-                ShaderCompiler::compile(expandedBuf.toUtf8(), QShader::FragmentStage);
-            errors += reportCompile(out, label, bufResult, declaredParamNames(eff.parameters));
-        }
-    }
-
-    // ── vertex stage ──
-    // The daemon warm-bake compiles the pack's declared vertexShader, falling back to
-    // shared/animation.vert (shader_warmup.cpp). Without baking it here a broken
-    // animation vert — or a regression in the shared vert every pack inherits —
-    // passes the shader_validate_bundled CI gate and only fails at runtime. The
-    // sibling zone and surface validators have always baked their vert stage.
-    if (PhosphorAnimationShaders::shaderEffectIsCompositorOnly(eff)) {
-        // A compositor-only pack's vert is the stage most likely to break: it
-        // is where the geometry-morph packs do their per-vertex work, and it
-        // is compiled on the KWin path ONLY, so nothing else in a headless run
-        // looks at it. Unlike the daemon vertex bake below, the p_<id> preamble
-        // IS spliced, because the compositor splices it for the vertex stage
-        // too (shader_textures.cpp) and vertex-driven packs read their params
-        // inside the PLASMAZONES_KWIN branch this bake takes.
-        if (!eff.vertexShaderPath.isEmpty() && QFile::exists(eff.vertexShaderPath)) {
-            errors += bakeCompositorStage(out, packDir, eff, eff.vertexShaderPath,
-                                          QFileInfo(eff.vertexShaderPath).fileName(), QStringLiteral("vert"),
-                                          /*scaffold=*/false);
-        }
-    } else {
-        const QString animIncludeDir = QFileInfo(packDir).absolutePath() + QStringLiteral("/shared");
-        QString vertPath = eff.vertexShaderPath;
-        if (vertPath.isEmpty()) {
-            const QString sharedVert = animIncludeDir + QStringLiteral("/animation.vert");
-            if (QFile::exists(sharedVert)) {
-                vertPath = sharedVert;
-            }
-        }
-        if (!vertPath.isEmpty() && QFile::exists(vertPath)) {
-            const QString vertLabel = QFileInfo(vertPath).fileName();
-            QFile vertFile(vertPath);
-            if (!vertFile.open(QIODevice::ReadOnly | QIODevice::Text)) {
-                out << "  " << vertLabel.leftJustified(15) << "ERROR\n    cannot read " << vertPath << "\n";
-                ++errors;
-            } else {
-                const QString rawVert = QString::fromUtf8(vertFile.readAll());
-                QString vertErr;
-                const QString expandedVert = ShaderCompiler::expandSource(rawVert, QFileInfo(vertPath).absolutePath(),
-                                                                          {animIncludeDir}, &vertErr);
-                if (expandedVert.isEmpty()) {
-                    out << "  " << vertLabel.leftJustified(15) << "ERROR\n    include expansion failed: " << vertErr
-                        << "\n";
-                    ++errors;
-                } else {
-                    // NO p_<id> preamble splice here: the runtime splices only
-                    // the FRAGMENT stage (loadFragmentShader and
-                    // warmShaderBakeCacheForPaths both do; loadVertexShader
-                    // does not). Splicing it here would let a daemon-path vert
-                    // that reads p_<id> pass this gate and then fail at runtime
-                    // with an undeclared identifier. Vertex-driven packs read
-                    // their params inside `#ifdef PLASMAZONES_KWIN`, which this
-                    // Vulkan-dialect bake does not take.
-                    const ShaderCompiler::Result vertResult =
-                        ShaderCompiler::compile(expandedVert.toUtf8(), QShader::VertexStage);
-                    errors += reportCompile(out, vertLabel, vertResult, declaredParamNames(eff.parameters));
-                }
-            }
-        }
-    }
-
-    if (errors == 0) {
-        out << "  → OK\n\n";
-    } else {
-        out << "  → " << errors << (errors == 1 ? " error\n\n" : " errors\n\n");
-    }
-    return errors;
-}
-
-// Validate one SURFACE pack directory (data/surface/*). Reproduces the surface
-// runtime's fragment assembly on the daemon (Qt-RHI) path — the pSurface entry
-// scaffold (an entry-only pack gets a generated main(); a pack with its own
-// main() passes through unchanged) + include expansion + the generated p_<id>
-// preamble — then bakes through headless glslang. Returns the error count.
-//
-// As with animation packs, the kwin-effect classic-GL branch
-// (`#define PLASMAZONES_KWIN`, default-block uniforms) is NOT baked here:
-// QShaderBaker compiles Vulkan-dialect GLSL and rejects default-block uniforms.
-// Baking the #else branch validates the daemon UBO contract in
-// surface_uniforms.glsl; the PLASMAZONES_KWIN plumbing is identical for every
-// pack and exercised by the live compositor compile.
-int validateSurfacePack(const QString& packDir, QTextStream& out)
-{
-    const QString name = QFileInfo(packDir).fileName();
-
-    QFile metaFile(QDir(packDir).filePath(QStringLiteral("metadata.json")));
-    if (!metaFile.open(QIODevice::ReadOnly)) {
-        out << name << "\n  metadata       ERROR\n    cannot read metadata.json\n  → 1 error\n\n";
-        return 1;
-    }
-    QJsonParseError perr{};
-    const QJsonDocument doc = QJsonDocument::fromJson(metaFile.readAll(), &perr);
-    if (doc.isNull() || !doc.isObject()) {
-        out << name << "\n  metadata       ERROR\n    invalid JSON: " << perr.errorString() << "\n  → 1 error\n\n";
-        return 1;
-    }
-
-    SurfaceShaderEffect eff = SurfaceShaderEffect::fromJson(doc.object());
-    eff.sourceDir = QDir(packDir).absolutePath();
-    // The `fragmentShader` / `bufferShaders` / `vertexShader` paths come from the
-    // user-editable metadata.json, so confine each to the pack dir (via
-    // confinedPackPath) before it is opened and fed to glslang: a
-    // `../../../etc/...` or absolute path must be rejected, not compiled. An
-    // empty path is left as-is (that stage is simply absent).
-    const auto confineToPack = [&packDir](QString& path) {
-        return confinePackPathInPlace(packDir, path);
-    };
-    if (!confineToPack(eff.fragmentShaderPath)) {
-        out << name
-            << "\n  metadata       ERROR\n    fragmentShader path escapes the pack directory (path traversal "
-               "rejected)\n  → 1 error\n\n";
-        return 1;
-    }
-    for (QString& b : eff.bufferShaderPaths) {
-        // `builtin:` tokens resolve against the surface shared/ dir (fixed
-        // whitelist, same resolver as the runtime registry) rather than the
-        // pack dir; an unknown token resolves empty and is linted below as a
-        // missing buffer, mirroring the runtime's fail-closed path.
-        if (SurfaceShaderRegistry::isBuiltinBufferShader(b)) {
-            b = SurfaceShaderRegistry::resolveBuiltinBufferShader(b, QDir(packDir).absolutePath());
-            continue;
-        }
-        if (!confineToPack(b)) {
-            out << name
-                << "\n  metadata       ERROR\n    bufferShaders path escapes the pack directory (path traversal "
-                   "rejected)\n  → 1 error\n\n";
-            return 1;
-        }
-    }
-    if (!confineToPack(eff.vertexShaderPath)) {
-        out << name
-            << "\n  metadata       ERROR\n    vertexShader path escapes the pack directory (path traversal rejected)\n "
-               " "
-               "→ 1 error\n\n";
-        return 1;
-    }
-    if (!eff.isValid()) {
-        out << name << "\n  metadata       ERROR\n    missing required field (id / fragmentShader)\n  → 1 error\n\n";
-        return 1;
-    }
-    const QString fragLabel = QFileInfo(eff.fragmentShaderPath).fileName();
-
-    out << name << "  (" << eff.parameters.size() << " param" << (eff.parameters.size() == 1 ? "" : "s") << ", "
-        << eff.textures.size() << " texture" << (eff.textures.size() == 1 ? "" : "s") << ", "
-        << (eff.isMultipass ? "multipass" : "single-pass") << ")\n";
-
-    int errors = 0;
-
-    // ── metadata lints ──
-    static const QStringList kSurfaceParamTypes = {QStringLiteral("float"), QStringLiteral("int"),
-                                                   QStringLiteral("bool"), QStringLiteral("color")};
-    QStringList lints;
-    for (const SurfaceShaderEffect::ParameterInfo& p : eff.parameters) {
-        if (!kSurfaceParamTypes.contains(p.type)) {
-            lints << QStringLiteral("unknown param type '%1' for '%2' (surface params are float/int/bool/color)")
-                         .arg(p.type, p.id);
-        }
-        if (!PhosphorShaders::isValidParamId(p.id)) {
-            lints
-                << QStringLiteral("invalid parameter id '%1' (not a GLSL identifier; skipped, no p_ define)").arg(p.id);
-        }
-    }
-    const QJsonArray declaredTextures = doc.object().value(QLatin1String("textures")).toArray();
-    if (declaredTextures.size() > PhosphorSurfaceShaders::SurfaceShaderContract::kMaxUserTextureSlots) {
-        lints << QStringLiteral("too many textures: %1 declared, cap is %2 (surplus dropped at load)")
-                     .arg(static_cast<int>(declaredTextures.size()))
-                     .arg(PhosphorSurfaceShaders::SurfaceShaderContract::kMaxUserTextureSlots);
-    }
-    for (const QJsonValue& v : declaredTextures) {
-        const QString texPath = v.toObject().value(QLatin1String("path")).toString();
-        if (texPath.isEmpty()) {
-            lints << QStringLiteral("texture entry with empty `path` (dropped at load)");
-        } else {
-            // Same confinement and existence check the animation arm applies,
-            // and for the same reason: the registry clears a rejected texture
-            // path and the sampler falls back to transparent, so a typo ships
-            // green and fails at first paint.
-            const auto confined = confinedPackPath(packDir, texPath);
-            if (!confined) {
-                lints << QStringLiteral(
-                             "texture path escapes the pack directory: %1 (rejected at load, sampler reads "
-                             "transparent)")
-                             .arg(texPath);
-            } else if (!QFile::exists(*confined)) {
-                lints << QStringLiteral("texture missing: %1 (sampler reads transparent at load)").arg(texPath);
-            }
-        }
-        // Wrap vocabulary lint — read RAW metadata: SurfaceShaderEffect::fromJson
-        // silently clears an invalid wrap to clamp, so a lint over the parsed
-        // eff.textures could never surface an author's typo. Mirror fromJson's
-        // {clamp,repeat,mirror} guard so a bad wrap fails the validator instead.
-        const QString wrap = v.toObject().value(QLatin1String("wrap")).toString();
-        if (!wrap.isEmpty() && !PhosphorSurfaceShaders::SurfaceShaderContract::isValidWrapToken(wrap)) {
-            lints
-                << QStringLiteral("texture wrap not in {clamp,repeat,mirror}: %1 (cleared to clamp at load)").arg(wrap);
-        }
-    }
-    // Multipass buffer lints — read RAW metadata, not the parsed struct: fromJson
-    // clamps bufferScale into [0.125, 1.0] and drops missing buffers, so a lint
-    // over the parsed values would hide author errors.
-    if (eff.isMultipass) {
-        const QJsonArray declaredBuffers = doc.object().value(QLatin1String("bufferShaders")).toArray();
-        for (const QJsonValue& v : declaredBuffers) {
-            const QString bufName = v.toString();
-            if (bufName.isEmpty()) {
-                // fromJson SKIPS an empty entry while bufferWraps and
-                // bufferFilters keep every entry in place, and those arrays are
-                // positionally aligned with this one — so one empty entry shifts
-                // every later pass's wrap and filter override by one, silently.
-                lints << QStringLiteral(
-                    "empty bufferShaders entry (dropped at load, which shifts the bufferWraps and bufferFilters "
-                    "alignment for every later pass)");
-                continue;
-            }
-            if (SurfaceShaderRegistry::isBuiltinBufferShader(bufName)) {
-                if (SurfaceShaderRegistry::resolveBuiltinBufferShader(bufName, QDir(packDir).absolutePath())
-                        .isEmpty()) {
-                    lints << QStringLiteral("unknown or unlocatable builtin buffer shader: %1").arg(bufName);
-                }
-                continue;
-            }
-            const auto confined = confinedPackPath(packDir, bufName);
-            if (!confined) {
-                lints << QStringLiteral("multipass buffer shader path escapes the pack directory: %1").arg(bufName);
-            } else if (!QFile::exists(*confined)) {
-                lints << QStringLiteral("multipass buffer shader missing: %1").arg(bufName);
-            }
-        }
-        // The runtime caps buffer passes and drops the surplus with only a
-        // journal warning, the same "runtime hid the author error" class the
-        // sibling arms lint.
-        if (declaredBuffers.size() > PhosphorSurfaceShaders::SurfaceShaderEffect::kMaxBufferPasses) {
-            lints << QStringLiteral("too many buffer shaders: %1 declared, cap is %2 (surplus dropped at load)")
-                         .arg(static_cast<int>(declaredBuffers.size()))
-                         .arg(PhosphorSurfaceShaders::SurfaceShaderEffect::kMaxBufferPasses);
-        }
-        // bufferWraps / bufferFilters are positionally aligned to bufferShaders
-        // and lossy in BOTH directions at load: a longer array is trimmed and a
-        // shorter one is padded with the single-value default, neither with a
-        // warning. Flag any mismatch, matching the animation arm, rather than
-        // surplus alone — a short array is the likelier authoring slip.
-        const auto lintBufferArrayLen = [&](QLatin1String key) {
-            const QJsonArray arr = doc.object().value(key).toArray();
-            if (!arr.isEmpty() && arr.size() != declaredBuffers.size()) {
-                lints << QStringLiteral(
-                             "%1 has %2 entries for %3 buffer shaders (aligned positionally; "
-                             "surplus dropped and missing entries fall back at load)")
-                             .arg(QString(key))
-                             .arg(static_cast<int>(arr.size()))
-                             .arg(static_cast<int>(declaredBuffers.size()));
-            }
-        };
-        lintBufferArrayLen(QLatin1String("bufferWraps"));
-        lintBufferArrayLen(QLatin1String("bufferFilters"));
-        // Vocabulary, on all four spellings. validatedWrap / validatedFilter
-        // clear an unrecognised token to empty with a journal warning only.
-        const auto lintSurfaceTokens = [&lints, &doc](QLatin1String key, bool wrap) {
-            for (const QJsonValue& v : doc.object().value(key).toArray()) {
-                const QString tok = v.toString();
-                const bool ok = wrap ? PhosphorSurfaceShaders::SurfaceShaderContract::isValidWrapToken(tok)
-                                     : PhosphorSurfaceShaders::SurfaceShaderContract::isValidFilterToken(tok);
-                if (!tok.isEmpty() && !ok) {
-                    lints << QStringLiteral("%1 value '%2' not in vocabulary (cleared at load)").arg(QString(key), tok);
-                }
-            }
-        };
-        lintSurfaceTokens(QLatin1String("bufferWraps"), true);
-        lintSurfaceTokens(QLatin1String("bufferFilters"), false);
-        const QString singleWrap = doc.object().value(QLatin1String("bufferWrap")).toString();
-        if (!singleWrap.isEmpty() && !PhosphorSurfaceShaders::SurfaceShaderContract::isValidWrapToken(singleWrap)) {
-            lints << QStringLiteral("bufferWrap value '%1' not in vocabulary (cleared at load)").arg(singleWrap);
-        }
-        const QString singleFilter = doc.object().value(QLatin1String("bufferFilter")).toString();
-        if (!singleFilter.isEmpty()
-            && !PhosphorSurfaceShaders::SurfaceShaderContract::isValidFilterToken(singleFilter)) {
-            lints << QStringLiteral("bufferFilter value '%1' not in vocabulary (cleared at load)").arg(singleFilter);
-        }
-        const double rawScale = doc.object().value(QLatin1String("bufferScale")).toDouble(1.0);
-        if (rawScale < PhosphorShaders::kMinBufferScale || rawScale > PhosphorShaders::kMaxBufferScale) {
-            lints << QStringLiteral("bufferScale out of range [%1, %2]: %3 (clamped at load)")
-                         .arg(PhosphorShaders::kMinBufferScale)
-                         .arg(PhosphorShaders::kMaxBufferScale)
-                         .arg(rawScale);
-        }
-    }
-    if (!QFile::exists(eff.fragmentShaderPath)) {
-        lints << QStringLiteral("fragment shader missing: %1").arg(fragLabel);
-    }
-    // An explicit per-pack `vertexShader` was resolved to absolute above; if the
-    // author typo'd the path the vertex stage below silently skips it (the
-    // exists() guard bows out with no diagnostic), so lint it here the same way
-    // the fragment stage is linted. An empty vertexShaderPath is the normal
-    // shared-surface.vert case and is not an error.
-    if (!eff.vertexShaderPath.isEmpty() && !QFile::exists(eff.vertexShaderPath)) {
-        lints << QStringLiteral("vertex shader missing: %1").arg(QFileInfo(eff.vertexShaderPath).fileName());
-    }
-
-    if (lints.isEmpty()) {
-        out << "  metadata       OK\n";
-    } else {
-        out << "  metadata       ERROR\n";
-        for (const QString& l : lints) {
-            out << "    " << l << "\n";
-            ++errors;
-        }
-    }
-
-    // ── stage compile (reproduce the daemon runtime fragment assembly) ──
-    if (QFile::exists(eff.fragmentShaderPath)) {
-        QFile frag(eff.fragmentShaderPath);
-        if (!frag.open(QIODevice::ReadOnly | QIODevice::Text)) {
-            out << "  " << fragLabel.leftJustified(15) << "ERROR\n    cannot read " << eff.fragmentShaderPath << "\n";
-            ++errors;
-        } else {
-            const QString raw = QString::fromUtf8(frag.readAll());
-            const QString surfacePacksRoot = QFileInfo(packDir).absolutePath();
-            // Match the runtime SurfaceShaderItem::surfaceIncludePaths(): each
-            // surface data dir contributes its `shared` subdir AND the dir
-            // itself, so a shader that resolves an include from the packs-root
-            // (not just `shared/`) bakes identically here and can't false-fail
-            // the gate. The sibling zone validator uses the same
-            // {root/shared, root} pair.
-            const QStringList includePaths = {surfacePacksRoot + QStringLiteral("/shared"), surfacePacksRoot};
-            QString err;
-            // Assemble an entry-only pack (a `vec4 pSurface(vec2 uv)` body, no
-            // main()) into a full TU before expansion, identical to the daemon /
-            // kwin paths, so `pSurface` packs validate. A main() pack passes
-            // through unchanged.
-            const QString assembled = PhosphorShaders::assembleEntryPoint(
-                raw, SurfaceShaderRegistry::surfaceEntryPrologue(), SurfaceShaderRegistry::surfaceEntryCandidates());
-            const QString expanded = ShaderCompiler::expandSource(
-                assembled, QFileInfo(eff.fragmentShaderPath).absolutePath(), includePaths, &err);
-            if (expanded.isEmpty()) {
-                out << "  " << fragLabel.leftJustified(15) << "ERROR\n    include expansion failed: " << err << "\n";
-                ++errors;
-            } else {
-                const QString spliced =
-                    PhosphorShaders::spliceAfterVersion(expanded, SurfaceShaderRegistry::paramPreamble(eff));
-                const ShaderCompiler::Result result = ShaderCompiler::compile(spliced.toUtf8(), QShader::FragmentStage);
-                errors += reportCompile(out, fragLabel, result, declaredParamNames(eff.parameters));
-            }
-        }
-    }
-
-    // ── multipass buffer passes ──
-    // Buffer passes carry their own main() (no entry scaffold, no param preamble)
-    // and bake on the daemon Qt-RHI path, same as overlay packs. The compositor
-    // runtime executes them via the GL-FBO chain; both share this source.
-    if (eff.isMultipass) {
-        const QString surfacePacksRoot = QFileInfo(packDir).absolutePath();
-        // Match the runtime SurfaceShaderItem::surfaceIncludePaths(): each surface
-        // data dir contributes its `shared` subdir AND the dir itself, so a shader
-        // that resolves an include from the packs-root (not just `shared/`) bakes
-        // identically here and can't false-fail the gate. The sibling zone
-        // validator uses the same {root/shared, root} pair.
-        const QStringList includePaths = {surfacePacksRoot + QStringLiteral("/shared"), surfacePacksRoot};
-        for (const QString& buf : eff.bufferShaderPaths) {
-            if (!QFile::exists(buf)) {
-                continue; // missing buffers already linted above
-            }
-            // A `builtin:` buffer resolved via the QStandardPaths fallback (user
-            // pack with no sibling shared/ dir) lives OUTSIDE surfacePacksRoot,
-            // where its angle-includes (surface_blur.glsl) would miss the
-            // pack-derived include paths. Append the buffer's own dir so the
-            // installed shared dir resolves; for bundled packs this duplicates
-            // an entry already in the list, which the resolver tolerates.
-            const QStringList bufferIncludePaths = QStringList(includePaths) << QFileInfo(buf).absolutePath();
-            const QString label = QFileInfo(buf).fileName();
-            QFile bufFile(buf);
-            if (!bufFile.open(QIODevice::ReadOnly | QIODevice::Text)) {
-                out << "  " << label.leftJustified(15) << "ERROR\n    cannot read " << buf << "\n";
-                ++errors;
-                continue;
-            }
-            const QString rawBuf = QString::fromUtf8(bufFile.readAll());
-            QString err;
-            const QString expanded =
-                ShaderCompiler::expandSource(rawBuf, QFileInfo(buf).absolutePath(), bufferIncludePaths, &err);
-            if (expanded.isEmpty()) {
-                out << "  " << label.leftJustified(15) << "ERROR\n    include expansion failed: " << err << "\n";
-                ++errors;
-            } else {
-                const ShaderCompiler::Result result =
-                    ShaderCompiler::compile(expanded.toUtf8(), QShader::FragmentStage);
-                errors += reportCompile(out, label, result, declaredParamNames(eff.parameters));
-            }
-        }
-    }
-
-    // ── vertex stage ──
-    // Mirror the daemon runtime (SurfaceShaderItem::updatePaintNode): an explicit
-    // per-pack `vertexShader` wins, else a per-pack `surface.vert` beside the
-    // fragment, else a shared `surface.vert` from the include paths. The vertex
-    // stage gets no scaffold and no param preamble — it ships its own main() (the
-    // fragment stage's pSurface scaffold does not apply here). Without this a
-    // malformed vertex stage passes the validator and only fails at the live
-    // daemon — the sibling zone path (validatePack) already bakes the vertex
-    // stage, so surface validation must too.
-    {
-        const QString surfacePacksRoot = QFileInfo(packDir).absolutePath();
-        // Match the runtime SurfaceShaderItem::surfaceIncludePaths(): each surface
-        // data dir contributes its `shared` subdir AND the dir itself, so a shader
-        // that resolves an include from the packs-root (not just `shared/`) bakes
-        // identically here and can't false-fail the gate. The sibling zone
-        // validator uses the same {root/shared, root} pair.
-        const QStringList includePaths = {surfacePacksRoot + QStringLiteral("/shared"), surfacePacksRoot};
-        QString vertPath = eff.vertexShaderPath;
-        if (vertPath.isEmpty()) {
-            // Beside the FRAGMENT (matching the daemon runtime and the comment
-            // above), not merely inside packDir — a nested fragmentShader path
-            // resolves its sibling surface.vert the same way at runtime.
-            const QString vertLocal =
-                QFileInfo(eff.fragmentShaderPath).absolutePath() + QStringLiteral("/surface.vert");
-            if (QFile::exists(vertLocal)) {
-                vertPath = vertLocal;
-            } else {
-                for (const QString& incDir : includePaths) {
-                    const QString candidate = incDir + QStringLiteral("/surface.vert");
-                    if (QFile::exists(candidate)) {
-                        vertPath = candidate;
-                        break;
-                    }
-                }
-            }
-        }
-        if (!vertPath.isEmpty() && QFile::exists(vertPath)) {
-            const QString label = QFileInfo(vertPath).fileName();
-            QFile vertFile(vertPath);
-            if (!vertFile.open(QIODevice::ReadOnly | QIODevice::Text)) {
-                out << "  " << label.leftJustified(15) << "ERROR\n    cannot read " << vertPath << "\n";
-                ++errors;
-            } else {
-                const QString rawVert = QString::fromUtf8(vertFile.readAll());
-                QString err;
-                const QString expanded =
-                    ShaderCompiler::expandSource(rawVert, QFileInfo(vertPath).absolutePath(), includePaths, &err);
-                if (expanded.isEmpty()) {
-                    out << "  " << label.leftJustified(15) << "ERROR\n    include expansion failed: " << err << "\n";
-                    ++errors;
-                } else {
-                    const ShaderCompiler::Result result =
-                        ShaderCompiler::compile(expanded.toUtf8(), QShader::VertexStage);
-                    errors += reportCompile(out, label, result, declaredParamNames(eff.parameters));
-                }
-            }
-        }
-    }
-
-    if (errors == 0) {
-        out << "  → OK\n\n";
-    } else {
-        out << "  → " << errors << (errors == 1 ? " error\n\n" : " errors\n\n");
-    }
-    return errors;
-}
-
-} // namespace PlasmaZones::ShaderValidate
